+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Prevent error when updating character as a player without filesystem browse permissions.
 - Divine Favor effect left in default damage.
 - Stones Endurance effect used a d10, instead of a d12.
+- Monster: Fathomer spell parsing.
 
 # 3.7.14